+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Artifacts](https://nexus3.inventage.com/#browse/search=version%3D%22???%22)
 
+### Removed
+
+- **BREAKING**: Removed static configuration `applications` ([PORTAL-2144](https://inventage-all.atlassian.net/browse/PORTAL-2144)).
+
 ## [9.5.0]-202506100904-135-d216a4a - 2025-06-10
 
 [Artifacts](https://nexus3.inventage.com/#browse/search=version%3D%229.5.0-202506100904-135-d216a4a%22)